fix: playwright config and setup
