+++ playwright.config.ts
@@ -3,10 +3,6 @@ import { devices } from '@playwright/test';
 const config = {
 	testDir: 'tests',
 	forbidOnly: !!process.env.CI,
-	webServer: {
-		command: 'npm run build && npm run preview',
-		port: 4173
-	},
 	projects: [
 		{
 			name: 'chromium',
@@ -14,18 +10,6 @@ const config = {
 				...devices['Desktop Chrome']
 			}
 		},
-		{
-			name: 'chromium',
-			use: {
-				...devices['Desktop Chrome']
-			}
-		},
-		{
-			name: 'firefox',
-			use: {
-				...devices['Desktop Firefox']
-			}
-		}
 	]
 };
