+++ src/lib/utils/dom-utils.ts
@@ -1,0 +1,3 @@
+export const testIds = {
+  navigation: "navigation"
+}
