+++ tests/contstants.ts
@@ -1,0 +1,3 @@
+export const mdnUrl = 'https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Intl';
+export const localBaseURL = 'http://localhost:5173'
+export const defaultPageUnderTest = '/';
